requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dotenv==0.9.9
Flask==3.1.0
flatbuffers==25.2.10
gTTS==2.5.4
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.3
onnxruntime==1.20.1
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pygame==2.6.1
pytesseract==0.3.13
python-dotenv==1.0.1
pyttsx3==2.98
requests==2.32.3
sympy==1.13.3
urllib3==2.3.0
Werkzeug==3.1.3
